Begin of code cleaning
