+++ includes/twenty_one_sh.h
@@ -6,7 +6,7 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/11 16:15:27 by jjaniec           #+#    #+#             */
-/*   Updated: 2018/09/19 20:12:28 by sbrucker         ###   ########.fr       */
+/*   Updated: 2018/09/20 19:48:07 by sbrucker         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,11 @@
 #  define VERBOSE_MODE 0
 # endif
 
-extern const size_t g_cant_begin_with[7];
+extern const size_t g_cant_begin_with[];
+extern const size_t	g_token_bypass[];
+extern const size_t	g_tokens[];
+extern const size_t	g_next_tokens[][4];
+extern void	(* const g_node_placer[])(t_ast *, t_ast *);
 
 void		subp_string(char **s);
 t_lexeme	*subp_lexeme(t_lexeme *lex);
