+++ tests/main.c
@@ -6,11 +6,18 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/15 13:51:41 by jjaniec           #+#    #+#             */
-/*   Updated: 2018/09/03 15:51:36 by sbrucker         ###   ########.fr       */
+/*   Updated: 2018/09/20 18:12:17 by sbrucker         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "tests.h"
+#include <time.h>
+
+void waitFor (unsigned int secs)
+{
+    unsigned int retTime = time(0) + secs;   // Get finishing time.
+    while (time(0) < retTime);               // Loop until it arrives.
+}
 
 int	main(int argc, char **argv, char **envp)
 {
